Crreate AbstractResponseException as base class for response exceptions.
